chore: use AtomicReferenceArray to keep track of worker threads
For the most part the array of worker threads is read-only, however it
may be updated from any thread when a worker thread transitions to
blocking or back.

Previously, a separate atomic variable was used to introduce memory
barriers when accessing the array. This is discarded in favour of
AtomicReferenceArray that specifically manages access to elements of the
array.

The `getWorkerThreads` method is replaced by a more limited
`getWorkerThread(index: Int)` method. This used by `WorkerThread` when
checking for blocked threads, and the new method can be used to
specifically access a single worker thread.