fix: use worker thread index at startup
Naming of active worker threads was inconsistent: when initially
created, worker threads were numbered by the pool's
`blockedWorkerThreadNamingIndex` in constructor code; if they later
switched back from blocking state they would use the worker thread
index.

`nameIndex` is renamed to `blockedWorkerIndex`, to make explicit the
link to the counter that originates it. This is still used when naming a
blocked worker, as well as for ranking `cachedThreads` in the pool.

With this change, the (current) worker index is used for naming both
newly constructed worker threads, and threads that switch back from
blocking.